overhaul cli
it is simultaneously easier to add more CLI options due to the magic of
macros, and easier for the user to get debug output.

instead of a single `--verbose` or `-v`, there are now several flags so
you can get exactly the output you need.

also, I fixed align_shift because it was irking me.

also also, codegen can now show disassembly as a verbose flag