Compile-time Generics!
Functions can now accept generic parameters. The arguments to these
parameters must be known at compile-time and can be used to compute the
types in both the header and body of the function.

```cpp add :: (comptime T: left: T, right: T) -> T { res : T = left +
right; res }

add(i32, 2, 3); ```

It used to be that each file had an associated set of types for each
expression in that file. This worked because the the same expression in
the same file would always correspond to the same type no matter the
function it was contained in. "Fqns" (fully qualified names) were used
to refer to globals and "Inferrables" were used to refer to either
globals or lambdas (things which could be typed by hir_ty as a singular
unit).

That system has been replaced by "NaiveLoc" and "ConcreteLoc."

NaiveLoc refers to either a global or lambda which has unknown generic
arguments e.g. `add(_, _, _)`, and ConcreteLoc refers to either a global
or lambda which has a known and concrete set of generic arguments e.g.
`add(i32, _, _)`.

hir works with NaiveLocs, and hir_ty works with ConcreteLocs. Instead of
each file being associated with a set of types, now each individual
ConcreteLoc is associated with a set of types. This allows the function
bodies of `add(i32, _, _)` and `add(u64, _, _)` to have two separate
maps which each contain different mappings of local variables,
expressions, etc. to types.

There are also now three function types: Concrete, NaivePolymorphic, and
FunctionPointer. Concrete function types store the regular information
along with the exact ConcreteLoc of the function. This is like Rust
where each function is its own function type. NaivePolymorphic functions
are functions which are associated with a NaiveLoc, which haven't yet
been called with a set of compile-time arguments, and which won't be
type-checked until they're called. Function pointers are what's produced
when you use the function type syntax `(a: i32, b: f32) -> i32`, and can
accept any non-polymorphic concrete function which has matching
parameter and return type.

When you call a naive polymorphic function, it will compile the correct
arguments and using them will mark the function to be re-evaluated using
those correct comptime arguments. The function will then be re-evaluated
and will become a Concrete function, at which point the call
type-checking will proceed as normal.